swap network filter and owner filter position
